requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
anyio=3.5.0=py39h06a4308_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py39h7f8727e_0
asttokens=2.0.5=pyhd3eb1b0_0
astunparse=1.6.3=pypi_0
attrs=22.1.0=py39h06a4308_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py39h06a4308_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.5=py39h7deecbd_0
brotli=1.0.9=h166bdaf_7
brotli-bin=1.0.9=h166bdaf_7
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.5.7=hbcca054_0
cachetools=5.3.0=pypi_0
certifi=2023.5.7=py39h06a4308_0
cffi=1.15.1=py39h5eee18b_3
charset-normalizer=3.1.0=pypi_0
comm=0.1.2=py39h06a4308_0
contourpy=1.0.5=py39hdb19cb5_0
cryptography=39.0.1=py39h9ce1e76_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cudatoolkit=11.8.0=h6a678d5_0
cycler=0.11.0=pyhd8ed1ab_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py39h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py39h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.2.10=h9c3ff4c_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py39h06a4308_0
flatbuffers=23.5.9=pypi_0
flit-core=3.6.0=pyhd3eb1b0_0
fontconfig=2.14.1=hef1e5e3_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=h0708190_1
gast=0.4.0=pypi_0
giflib=5.2.1=h36c2ea0_2
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-auth=2.18.1=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.54.2=pypi_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
h5py=3.8.0=pypi_0
icu=58.2=he6710b0_3
idna=3.4=py39h06a4308_0
importlib-metadata=6.6.0=pypi_0
importlib_resources=5.12.0=pyhd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.19.2=py39hb070fc8_0
ipython=8.8.0=py39h06a4308_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jax=0.4.10=pypi_0
jedi=0.18.1=py39h06a4308_1
jinja2=3.1.2=py39h06a4308_0
jpeg=9e=h166bdaf_1
jsonschema=4.16.0=py39h06a4308_0
jupyter_client=7.4.8=py39h06a4308_0
jupyter_core=5.1.1=py39h06a4308_0
jupyter_server=1.23.4=py39h06a4308_0
jupyterlab_pygments=0.1.2=py_0
keras=2.12.0=pypi_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.4=py39h6a678d5_0
krb5=1.19.3=h3790be6_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h166bdaf_7
libbrotlidec=1.0.9=h166bdaf_7
libbrotlienc=1.0.9=h166bdaf_7
libclang=16.0.0=pypi_0
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.6.1.9=0
libcurand=10.3.2.106=0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h5eee18b_0
libedit=3.1.20191231=he28a2e2_2
libevent=2.1.12=h8f2d780_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libllvm10=10.0.1=he513fc3_3
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libpng=1.6.39=h5eee18b_0
libpq=12.9=h16c4e8d_3
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
lxml=4.9.1=py39h1edc446_0
lz4-c=1.9.3=h9c3ff4c_1
markdown=3.4.3=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.7.1=py39hf3d152e_0
matplotlib-base=3.7.1=py39h417a72b_1
matplotlib-inline=0.1.6=py39h06a4308_0
mistune=0.8.4=py39h27cfd23_1000
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
ml-dtypes=0.1.0=pypi_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.2.1=py39h06a4308_0
munkres=1.1.4=pyh9f0ad1d_0
nbclassic=0.4.8=py39h06a4308_0
nbclient=0.5.13=py39h06a4308_0
nbconvert=6.5.4=py39h06a4308_0
nbformat=5.7.0=py39h06a4308_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=py39h06a4308_0
nettle=3.7.3=hbbd107a_1
networkx=2.8.4=py39h06a4308_1
notebook=6.5.2=py39h06a4308_0
notebook-shim=0.2.2=py39h06a4308_0
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numexpr=2.8.4=py39he184ba9_0
numpy=1.23.5=py39h14f4228_0
numpy-base=1.23.5=py39h31eccc5_0
nvidia-cublas-cu11=11.11.3.6=pypi_0
nvidia-cuda-runtime-cu11=11.8.89=pypi_0
nvidia-cudnn-cu11=8.6.0.163=pypi_0
oauthlib=3.2.2=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1t=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=23.1=pypi_0
pandas=1.5.3=py39h417a72b_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h9c3ff4c_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.4.0=py39h6a678d5_0
pip=23.1.2=pypi_0
platformdirs=2.5.2=py39h06a4308_0
plotly=5.14.1=py_0
ply=3.11=py_1
prometheus_client=0.14.1=py39h06a4308_0
prompt-toolkit=3.0.36=py39h06a4308_0
protobuf=4.23.1=pypi_0
psutil=5.9.0=py39h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=23.0.0=py39h06a4308_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.15.7=py39h6a678d5_1
pyqt5-sip=12.11.0=py39h6a678d5_1
pyrsistent=0.18.0=py39heee7806_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.16=h7a1cb2a_2
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py39h06a4308_0
python_abi=3.9=2_cp39
pytorch=2.0.1=py3.9_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_5
pytorch-mutex=1.0=cuda
pytz=2022.7=py39h06a4308_0
pyzmq=23.2.0=py39h6a678d5_0
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
readline=8.2=h5eee18b_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scipy=1.10.1=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=66.0.0=py39h06a4308_0
sip=6.6.2=py39h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.2.0=py39h06a4308_1
soupsieve=2.3.2.post1=py39h06a4308_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.2.0=pyhd3eb1b0_0
sympy=1.11.1=py39h06a4308_0
tenacity=8.2.2=py39h06a4308_0
tensorboard=2.12.3=pypi_0
tensorboard-data-server=0.7.0=pypi_0
tensorflow=2.12.0=pypi_0
tensorflow-estimator=2.12.0=pypi_0
tensorflow-io-gcs-filesystem=0.32.0=pypi_0
tensorrt=8.5.3.1=pypi_0
termcolor=2.3.0=pypi_0
terminado=0.17.1=py39h06a4308_0
tinycss2=1.2.1=py39h06a4308_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd8ed1ab_0
torchaudio=2.0.2=py39_cu117
torchtriton=2.0.0=py39
torchvision=0.15.2=py39_cu117
tornado=6.2=py39h5eee18b_0
traitlets=5.7.1=py39h06a4308_0
typing-extensions=4.5.0=pypi_0
typing_extensions=4.4.0=py39h06a4308_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.15=py39h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39h06a4308_1
websocket-client=0.58.0=py39h06a4308_4
werkzeug=2.3.4=pypi_0
wheel=0.38.4=py39h06a4308_0
wrapt=1.14.1=pypi_0
xz=5.4.2=h5eee18b_0
zeromq=4.3.4=h2531618_0
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
